refactor: rename component files and Product -> ProductCard
