+++ client/components/ProductCard.spec.js
@@ -4,17 +4,17 @@ import {expect} from 'chai'
 import React from 'react'
 import enzyme, {shallow} from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import {Product} from './product'
+import {ProductCard} from './ProductCard'
 
 const adapter = new Adapter()
 enzyme.configure({adapter})
 
-describe('Product component', () => {
+describe('Product Cardcomponent', () => {
   let wrapper
   const product = {name: 'Fake glitter', price: 100}
 
   beforeEach(() => {
-    wrapper = shallow(<Product {...product} />)
+    wrapper = shallow(<ProductCard {...product} />)
   })
 
   it('renders the product name in an h4', () => {
