+++ client/components/index.js
@@ -3,8 +3,8 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {default as AllProducts} from './all-products'
-export {default as Product} from './product'
-export {Login, Signup} from './auth-form'
+export {default as Navbar} from './Navbar'
+export {default as UserHome} from './UserHome'
+export {default as AllProducts} from './AllProducts'
+export {default as ProductCard} from './ProductCard'
+export {Login, Signup} from './AuthForm'
